ui.ToolNode is now not set to None every OnUpdate
Previously if the tool was not enabled, every loop ui.toolMode was set
to None.
Inside the ModUI.toolMode setter the selected index of the
builtinTabstrip is set to the previous value if the mode is None. Since
this setter was called every Update, every Update reset the selected
index to a value which visually selected a tab button.
This prevented all other mods from setting the selected Index to -1 and
therefore visually disabling the selected tab.

This fix resets the selected index only if the ToolMode is not None.